requirements.txt
django
gunicorn
django-heroku
django-cors-headers
requests
djangorestframework
scikit-surprise
pandas
